// ManShirtShopFE/man-shirt-admin/src/app/core/models/contact.models.ts
export class contactResponse{
  id : number;
  name : string;
  addressName : string;
  district : string;
  phone: string;
  province: string;
  ward : string;
  status: number;
  createTime: Date;
  updateTime: Date;
  createBy: string;
  updateBy: string;
}
export class contactRequest{
  id : number;
  name : string;
  addressName : string;
  district : string;
  phone: string;
  province: string;
  ward : string;
  status: number;
  createTime: Date;
  updateTime: Date;
  createBy: string;
  updateBy: string;
}